api_service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
pymongo==4.9.2
python-dotenv==1.0.1
bcrypt==4.2.1
pydantic==2.10.3
email-validator==2.3.0
requests==2.32.3
apscheduler==3.10.4
pydantic-settings==2.6.1
